third_party/blink/web_tests/wpt_internal/css/fit-width-text/text-grow-valid.html

<!DOCTYPE html>
<title>Parsing text-fit with valid values</title>
<link rel="help" href="https://github.com/explainers-by-googlers/css-fit-text/blob/main/README.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_valid_value('text-fit', 'none');
test_valid_value('text-fit', 'grow');
test_valid_value('text-fit', 'shrink');

test_valid_value('text-fit', 'grow consistent', 'grow consistent');
test_valid_value('text-fit', 'grow per-line');
test_valid_value('text-fit', 'grow per-line-all');

// TODO(crbug.com/417306102): The `limit` component should be a percentage value.
test_valid_value('text-fit', 'grow per-line 32px');
test_valid_value('text-fit', 'grow consistent 3em');

test_valid_value('text-fit', 'shrink per-line 4ch', 'shrink per-line 4ch');
test_valid_value('text-fit', 'shrink consistent 2lh', 'shrink consistent 2lh');
</script>
</body>
